requirements.txt
asgiref==3.7.2
astroid==3.1.0
cfgv==3.4.0
colorama==0.4.6
dill==0.3.8
distlib==0.3.8
Django==5.0.3
django-debug-toolbar==4.3.0
filelock==3.13.1
identify==2.5.35
isort==5.13.2
mccabe==0.7.0
nodeenv==1.8.0
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.6.2
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
pylint==3.1.0
python-dotenv==1.0.1
PyYAML==6.0.1
sqlparse==0.4.4
tomlkit==0.12.4
typing_extensions==4.10.0
tzdata==2024.1
virtualenv==20.25.1
